test(query-core): align mutation test filenames with 'mutation.ts' by renaming and merging duplicate '.test-d' files
